// src/app/core/models/Product.ts
export interface Product {
    title: string;
    category: string;
    images: string[];
    brand: string;
    price: number;
    cpu: string;
    camera: string;
    size: string;
    weight: string;
    display: string;
    battery: string;
    memory: string;
    id: number;
    quantity?: number;
    description: string;
  }